To try out
I thought I had samples working, but alas not. It is playing a sound, but it is not correct. Pressing 4 should play the sound "letsgo" ie, sound like the 'let's go' in lemmings. But, it doesn't, more like a bird chirp. I might be doing something wrong with the array, maybe. The data seems correct as it is the same binary size as the original raw sound file which I have used on my diamonds game.